app/requirements.txt
bcrypt==3.2.0
cffi==1.14.4
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
Flask-Testing==0.8.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2-binary==2.8.6
pycparser==2.20
python-dotenv==0.15.0
six==1.15.0
SQLAlchemy==1.3.22
Werkzeug==1.0.1
